+++ clients/tabby-playground/.env.example
@@ -1,20 +1,1 @@
-# You must first activate a Billing Account here: https://platform.openai.com/account/billing/overview
-# Then get your OpenAI API Key here: https://platform.openai.com/account/api-keys
-OPENAI_API_KEY=XXXXXXXX
-
-# Generate a random secret: https://generate-secret.vercel.app/32 or `openssl rand -base64 32`
-AUTH_SECRET=XXXXXXXX
-# Create a GitHub OAuth app here: https://github.com/settings/applications/new
-# Authorization callback URL: https://authjs.dev/reference/core/providers_github#callback-url
-AUTH_GITHUB_ID=XXXXXXXX
-AUTH_GITHUB_SECRET=XXXXXXXX
-# Support OAuth login on preview deployments, see: https://authjs.dev/guides/basics/deployment#securing-a-preview-deployment
-# Set the following only when deployed. In this example, we can reuse the same OAuth app, but if you are storing users, we recommend using a different OAuth app for development/production so that you don't mix your test and production user base.
-# AUTH_REDIRECT_PROXY_URL=https://YOURAPP.vercel.app/api/auth
-
-# Instructions to create kv database here: https://vercel.com/docs/storage/vercel-kv/quickstart and 
-KV_URL=XXXXXXXX
-KV_REST_API_URL=XXXXXXXX
-KV_REST_API_TOKEN=XXXXXXXX
-KV_REST_API_READ_ONLY_TOKEN=XXXXXXXX
-
+NEXT_PUBLIC_TABBY_SERVER_URL=http://127.0.0.1:8080
