+++ clients/tabby-playground/components/chat.tsx
@@ -8,14 +8,16 @@ import { ChatPanel } from '@/components/chat-panel'
 import { EmptyScreen } from '@/components/empty-screen'
 import { ChatScrollAnchor } from '@/components/chat-scroll-anchor'
 import { toast } from 'react-hot-toast'
+import { usePatchFetch } from "@/lib/hooks/use-patch-fetch"
 
-const IS_PREVIEW = process.env.VERCEL_ENV === 'preview'
 export interface ChatProps extends React.ComponentProps<'div'> {
   initialMessages?: Message[]
   id?: string
 }
 
 export function Chat({ id, initialMessages, className }: ChatProps) {
+  usePatchFetch()
+  
   const { messages, append, reload, stop, isLoading, input, setInput } =
     useChat({
       initialMessages,
