+++ clients/tabby-playground/next.config.js
@@ -1,6 +1,7 @@
 /** @type {import('next').NextConfig} */
 module.exports = {
   reactStrictMode: true,
+  output: "export",
   images: {
     unoptimized: true
   },
